% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{flip_coin}
\alias{flip_coin}
\title{Flip a fair coin}
\usage{
flip_coin(flips = 100L, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{flips}{an integer (defaults to 100)}

\item{seed}{random number seed}
}
\value{
\code{ggplot} object
}
\description{
Useful for demonstrating the Law of Large Numbers
}
\examples{
# Flip a coin 1000 times and plot the probabilities of Heads and Tails:
flip_coin(flips = 1000)
}
\author{
Lawrence R. De Geest
}
