% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hprice}
\alias{hprice}
\title{Housing prices and attributes from 1976.}
\format{
A data frame with 506 rows and 12 variables, including:
\describe{
  \item{lprice}{log median house price in the community}
  \item{nox}{Nitrous Oxide (NOX) concentration; parts per million}
  \item{crime}{number of reported crimes per capita}
  \item{rooms}{average number of rooms in houses in the community}
  \item{dist}{weighted distance of the community to 5 employment centers}
  \item{stratio}{average student-teacher ratio of schools in the community}
  \item{proptax}{property taxes}
  ...
}
}
\source{
Harrison, Jr., David, Rubinfeld, Daniel L. (1978/03)."Hedonic housing prices and the demand for clean air." Journal of Environmental Economics and Management 5(1): 81-102. <http://hdl.handle.net/2027.42/22636> \url{https://deepblue.lib.umich.edu/handle/2027.42/22636}
}
\usage{
hprice
}
\description{
Housing prices and attributes from 1976.
}
\note{
\code{hprice} is useful for teaching hedonic pricing (namely, the relative value of pollution as captured by housing prices), and linear regression in general.
}
\keyword{datasets}
