% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Body measurements for three species of penguins on the Palmer Archipelago in Antarctica.}
\format{
A data frame with 333 rows and 8 variables:
\describe{
  \item{species}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
  \item{island}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
  \item{bill_length_mm}{a number denoting bill length (millimeters)}
  \item{bill_depth_mm}{a number denoting bill depth (millimeters)}
  \item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
  \item{body_mass_g}{an integer denoting body mass (grams)}
  \item{sex}{a factor denoting penguin sex (female, male)}
  \item{year}{an integer denoting the study year (2007, 2008, or 2009)}
}
}
\source{
This data set is borrowed from the R package \code{palmerpenguins} by Allison Horst, Alison Hill and Kristen Gorman (\url{https://allisonhorst.github.io/palmerpenguins/}). Check out their webpage for complimentary penguin artwork!
}
\usage{
penguins
}
\description{
Body measurements for three species of penguins on the Palmer Archipelago in Antarctica.
}
\note{
\code{penguins} is a great introductory data set used throughout the data science and machine learning community (all the variables are exogenous; penguins can't choose their measurements!). To make things a bit easier, all rows with NAs were removed (to avoid issues like passing \code{na.rm = TRUE} to summary routines like \code{mean()} and so on).
}
\keyword{datasets}
