% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_probability}
\alias{plot_probability}
\title{Plot probabilities as areas underneath the PDF}
\usage{
plot_probability(
  model,
  lower = NULL,
  upper = NULL,
  distribution = NULL,
  details = NULL,
  fill = "grey35",
  color = "grey35",
  alpha = 1,
  ...
)
}
\arguments{
\item{model}{one of R's built-in probability distributions (e.g., 'rnorm', 'rpois', 'runif').}

\item{lower}{lower bound of P(a <= x <= b), i.e. b. Defaults to NULL.}

\item{details}{details about the distribution (e.g., mean and sd). Used in the plot subtitle(s). Defaults to NULL.}

\item{fill}{fill color of the histogram or density. Defaults to "grey35" (the ggplot2 default)}

\item{alpha}{transparency of the histogram or density. Defaults to 1 (no transparency)}

\item{...}{additional arguments passed to `model` (e.g. mean and sd for `rnorm`, lambda for `rpois`, etc.)}

\item{uppder}{upper bound of P(a <= x <= b), i.e. a. Defaults to NULL.}

\item{density}{plot the kernel density estimates instead of the histogram. Defaults to FALSE.}

\item{name}{name of the distribution (a string). Used in the plot title(s). Defaults to NULL, in which case `model` is used in the plot title.}
}
\value{
\code{ggplot} object
}
\description{
Plot probabilities as areas underneath the PDF
}
\note{
Should this function merge with `plot_distribution()`?
}
\examples{
# plot the standard normal and show the entire distribution sums to 1
plot_probability(model = 'rnorm', lower = -Inf, distribution = "Standard Normal", details = "Mean 0 SD 1")

# ...and half the distribution sums to 0.5
plot_probability(model = 'rnorm', lower = 0, distribution = "Standard Normal", details = "Mean 0 SD 1")

# show the area between 0 and 1 on the standard normal
plot_probability(model = 'rnorm', lower = 0, upper =1,  distribution = "Standard Normal", details = "Mean 0 SD 1")

# Chi-squared distribution with 5 degrees of freedome
plot_probability(model = 'rchisq', df = 5, lower = 8, distribution = "Chi-squared", details = "5 degrees of freedom")
}
\author{
Lawrence R. De Geest
}
