% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_t_value}
\alias{plot_t_value}
\title{Plot t-distribution and visualize two-tailed p-value for a given t-statistic and degrees of freedom}
\usage{
plot_t_value(t_value, dof, fill = "grey35", color = "grey35", alpha = 1)
}
\arguments{
\item{dof}{degrees of freedom}

\item{fill}{fill of the shaded area underneath the distribution for a given t-value. Defaults to "grey35" (ggplot2 default)}

\item{color}{color of the outline of the kernel density. Defaults to "grey35" (ggplot2 default)}

\item{t-value}{a value from the Student's T distribution}
}
\value{
\code{ggplot} object
}
\description{
Plot t-distribution and visualize two-tailed p-value for a given t-statistic and degrees of freedom
}
\note{
add switch to plot standard normal and calculate p-values instead of t-distribution
}
\examples{
# plot t distribution for 100 degrees of freedom, then calculate and shade-in the p-value for t = 1.25
plot_t_value(t_value = 1.25, dof = 100)
}
\author{
Lawrence R. De Geest
}
