% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{roll_die}
\alias{roll_die}
\title{Roll a fair die}
\usage{
roll_die(rolls = 100L, seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{rolls}{an integer (defaults to 100)}

\item{seed}{random number seed}
}
\value{
\code{ggplot} object
}
\description{
Useful for demonstrating the Law of Large Numbers
}
\examples{
# Roll a die 1000 times and plot the probabilities of each outcome:
roll_die(rolls = 1000)
}
\author{
Lawrence R. De Geest
}
